Wrote Basic Max Money Min Security Script
- Wrote maxOutServer so it grows to max money while min security with some allowed leeway.
- Fixed purchaseMaxServer script.
- Modified purchaseServer script.
- Fixed purchaseservercost script.
- Fixed runScriptXTimes script so it runs on host machine the file is located on.

+++ purchaseMaxServer.js
@@ -1,9 +1,11 @@
 /** @param {NS} ns */
 export async function main(ns) {
-const server = ns.args[0];
-const ram = ns.args[1];
-await ns.killall(server);
-await ns.deleteServer(server);
-await ns.purchaseServer(server, ram);
-await ns.sleep(15000);
+	const server = ns.args[0];
+	const ram = ns.args[1];
+	if (ns.serverExists(server)) {
+		await ns.killall(server);
+		await ns.deleteServer(server);
+	}
+	await ns.purchaseServer(server, ram);
+
 }
